// src/models/syntax-tree-interfaces.ts
export interface VisibleEndpoint {
    kind?: string;
    isCaller: boolean;
    isExternal: boolean;
    isModuleVar: boolean;
    moduleName: string;
    name: string;
    orgName: string;
    typeName: string;
    viewState?: any;
  }
  
  export interface NodePosition {
    startLine?: number;
    startColumn?: number;
    endLine?: number;
    endColumn?: number;
  }
  
  export interface ControlFlow {
    isReached?: boolean;
    isCompleted?: boolean;
    numberOfIterations?: number;
    executionTime?: number;
  }
  
  export interface STNode {
    kind: string;
    value?: any;
    parent?: STNode;
    viewState?: any;
    dataMapperViewState?: any;
    dataMapperTypeDescNode?: STNode;
    position?: any;
    typeData?: any;
    VisibleEndpoints?: VisibleEndpoint[];
    source: string;
    configurablePosition?: NodePosition;
    controlFlow?: ControlFlow;
  }
// ...
  export interface AsteriskToken extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface BinaryExpression extends STNode {
    lhsExpr:
      | BinaryExpression
      | FieldAccess
      | FunctionCall
      | IndexedExpression
      | MethodCall
      | NumericLiteral
      | OptionalFieldAccess
      | QualifiedNameReference
      | SimpleNameReference
      | StringLiteral
      | OptionalFieldAccess;
    operator:
      | AsteriskToken
      | BitwiseAndToken
      | BitwiseXorToken
      | DoubleDotLtToken
      | DoubleEqualToken
      | EllipsisToken
      | ElvisToken
      | GtEqualToken
      | GtToken
      | LogicalAndToken
      | LogicalOrToken
      | LtEqualToken
      | LtToken
      | NotDoubleEqualToken
      | NotEqualToken
      | PercentToken
      | PipeToken
      | PlusToken
      | SlashToken
      | TrippleEqualToken;
    rhsExpr:
      | BinaryExpression
      | BracedExpression
      | CheckExpression
      | FieldAccess
      | FunctionCall
      | IndexedExpression
      | MethodCall
      | NilLiteral
      | NumericLiteral
      | QualifiedNameReference
      | SimpleNameReference
      | StringLiteral
      | StringTypeDesc
      | TypeCastExpression
      | TypeTestExpression;
  }
  
  export interface BitwiseAndToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface BitwiseXorToken extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface BooleanLiteral extends STNode {
    literalToken: FalseKeyword | TrueKeyword;
  }
  
  export interface BooleanTypeDesc extends STNode {
    name: BooleanKeyword;
  }
  
  export interface BracedExpression extends STNode {
    closeParen: CloseParenToken;
    expression:
      | BinaryExpression
      | BooleanLiteral
      | CheckExpression
      | FunctionCall
      | MethodCall
      | SimpleNameReference
      | TypeCastExpression
      | TypeTestExpression
      | UnaryExpression
      | XmlStepExpression;
    openParen: OpenParenToken;
  }
// ...
  export interface CheckAction extends STNode {
    checkKeyword: CheckKeyword | CheckpanicKeyword;
    expression: RemoteMethodCallAction;
  }
  
  export interface CheckExpression extends STNode {
    checkKeyword: CheckKeyword | CheckpanicKeyword;
    expression:
      | FieldAccess
      | FunctionCall
      | ImplicitNewExpression
      | MethodCall
      | SimpleNameReference;
  }
  
  export interface CheckKeyword extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface CheckpanicKeyword extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface CloseBracketToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface CloseParenToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface ColonToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface CommaToken extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface DecimalFloatingPointLiteralToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface DecimalIntegerLiteralToken extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface DotToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface DoubleDotLtToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface DoubleEqualToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface DoubleGtToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface DoubleLtToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface DoubleQuoteToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface EllipsisToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface ElseBlock extends STNode {
    elseBody: BlockStatement | IfElseStatement;
    elseKeyword: ElseKeyword;
  }
  
  export interface ElseKeyword extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface ElvisToken extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface ErrorConstructor extends STNode {
    arguments: (CommaToken | NamedArg | PositionalArg)[];
    closeParenToken: CloseParenToken;
    errorKeyword: ErrorKeyword;
    openParenToken: OpenParenToken;
    typeReference?: SimpleNameReference;
  }
// ...
  export interface ErrorTypeDesc extends STNode {
    errorKeywordToken: ErrorKeyword;
    errorTypeParamsNode?: ErrorTypeParams;
    name?: ErrorKeyword;
  }
// ...
  export interface FieldAccess extends STNode {
    dotToken: DotToken;
    expression:
      | FieldAccess
      | FunctionCall
      | IndexedExpression
      | MethodCall
      | SimpleNameReference;
    fieldName: SimpleNameReference;
  }
// ...
  export interface FunctionCall extends STNode {
    arguments: (CommaToken | NamedArg | PositionalArg | RestArg)[];
    closeParenToken: CloseParenToken;
    functionName: ErrorTypeDesc | QualifiedNameReference | SimpleNameReference;
    openParenToken: OpenParenToken;
  }
// ...
  export interface GtEqualToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface GtToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface HandleKeyword extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface HandleTypeDesc extends STNode {
    name: HandleKeyword;
  }
  
  export interface HashToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface IdentifierToken extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface ImplicitNewExpression extends STNode {
    newKeyword: NewKeyword;
    parenthesizedArgList?: ParenthesizedArgList;
  }
// ...
  export interface IndexedExpression extends STNode {
    closeBracket: CloseBracketToken;
    containerExpression:
      | FieldAccess
      | IndexedExpression
      | MethodCall
      | SimpleNameReference;
    keyExpression: (
      | FieldAccess
      | MethodCall
      | NumericLiteral
      | SimpleNameReference
      | StringLiteral
    )[];
    openBracket: OpenBracketToken;
  }
// ...
  export interface IsKeyword extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface LogicalAndToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface LogicalOrToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface LtEqualToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface LtToken extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface MethodCall extends STNode {
    arguments: (CommaToken | NamedArg | PositionalArg | RestArg)[];
    closeParenToken: CloseParenToken;
    dotToken: DotToken;
    expression:
      | BracedExpression
      | FieldAccess
      | FunctionCall
      | IndexedExpression
      | MethodCall
      | OptionalFieldAccess
      | SimpleNameReference
      | StringLiteral;
    methodName: SimpleNameReference;
    openParenToken: OpenParenToken;
  }
// ...
  export interface NamedArg extends STNode {
    argumentName: SimpleNameReference;
    equalsToken: EqualToken;
    expression:
      | BinaryExpression
      | BooleanLiteral
      | FunctionCall
      | MappingConstructor
      | NumericLiteral
      | QualifiedNameReference
      | SimpleNameReference
      | StringLiteral
      | StringTypeDesc;
  }
// ...
  export interface NilLiteral extends STNode {
    closeParenToken: CloseParenToken;
    openParenToken: OpenParenToken;
  }
  
  export interface NilTypeDesc extends STNode {
    closeParenToken: CloseParenToken;
    openParenToken: OpenParenToken;
  }
  
  export interface NotDoubleEqualToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface NotEqualToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface NullKeyword extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface NullLiteral extends STNode {
    literalToken: NullKeyword;
    source: string;
  }
  
  export interface NumericLiteral extends STNode {
    literalToken: DecimalFloatingPointLiteralToken | DecimalIntegerLiteralToken;
  }
// ...
  export interface OpenBracketToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface OpenParenToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface OptionalChainingToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface OptionalFieldAccess extends STNode {
    expression: SimpleNameReference | OptionalFieldAccess | MethodCall;
    fieldName: SimpleNameReference;
    optionalChainingToken: OptionalChainingToken;
  }
// ...
  export interface PercentToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface PipeToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface PlusToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface PositionalArg extends STNode {
    expression:
      | BinaryExpression
      | BooleanLiteral
      | CheckExpression
      | ExplicitAnonymousFunctionExpression
      | FieldAccess
      | FunctionCall
      | ImplicitAnonymousFunctionExpression
      | IndexedExpression
      | IntTypeDesc
      | JsonTypeDesc
      | MappingConstructor
      | MethodCall
      | NilLiteral
      | NumericLiteral
      | OptionalFieldAccess
      | QualifiedNameReference
      | SimpleNameReference
      | StringLiteral
      | StringTemplateExpression
      | TypeCastExpression
      | TypeTestExpression
      | UnaryExpression
      | XmlTemplateExpression
      | ListConstructor;
  }
  
  export interface PrivateKeyword extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface PublicKeyword extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface QualifiedNameReference extends STNode {
    colon: ColonToken;
    identifier: IdentifierToken;
    modulePrefix: IdentifierToken;
  }
// ...
  export interface RemoteMethodCallAction extends STNode {
    arguments: (CommaToken | NamedArg | PositionalArg)[];
    closeParenToken: CloseParenToken;
    expression: FieldAccess | SimpleNameReference;
    methodName: SimpleNameReference;
    openParenToken: OpenParenToken;
    rightArrowToken: RightArrowToken;
  }
// ...
  export interface RestArg extends STNode {
    ellipsis: EllipsisToken;
    expression: SimpleNameReference;
  }
// ...
  export interface SimpleNameReference extends STNode {
    name: IdentifierToken;
  }
  
  export interface SlashLtToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface SlashToken extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface StringKeyword extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface StringLiteral extends STNode {
    literalToken: StringLiteralToken;
  }
  
  export interface StringLiteralToken extends STNode {
    isToken: boolean;
    value: string;
  }
  
  export interface StringTemplateExpression extends STNode {
    content: (Interpolation | TemplateString)[];
    endBacktick: BacktickToken;
    startBacktick: BacktickToken;
    type: StringKeyword;
  }
  
  export interface StringTypeDesc extends STNode {
    name: StringKeyword;
  }
// ...
  export interface TrippleEqualToken extends STNode {
    isToken: boolean;
    value: string;
  }
// ...
  export interface TypeCastExpression extends STNode {
    expression:
      | BracedExpression
      | CheckAction
      | CheckExpression
      | ErrorConstructor
      | FieldAccess
      | FunctionCall
      | IndexedExpression
      | MethodCall
      | SimpleNameReference
      | TypeCastExpression
      | RemoteMethodCallAction;
    gtToken: GtToken;
    ltToken: LtToken;
    typeCastParam: TypeCastParam;
  }
// ...
  export interface TypeTestExpression extends STNode {
    expression: SimpleNameReference;
    isKeyword: IsKeyword;
    typeDescriptor:
      | ArrayTypeDesc
      | BooleanTypeDesc
      | ErrorTypeDesc
      | FloatTypeDesc
      | IntTypeDesc
      | JsonTypeDesc
      | NilTypeDesc
      | ParameterizedTypeDesc
      | QualifiedNameReference
      | RecordTypeDesc
      | SimpleNameReference
      | StreamTypeDesc
      | StringTypeDesc
      | TableTypeDesc
      | XmlTypeDesc;
  }
// ...
  export interface UnaryExpression extends STNode {
    expression:
      | BracedExpression
      | MethodCall
      | NumericLiteral
      | SimpleNameReference;
    unaryOperator: ExclamationMarkToken | MinusToken;
  }
// ...
  export interface XmlStepExpression extends STNode {
    expression: SimpleNameReference;
    xmlStepStart: XmlNamePatternChain;
  }